TMS: Trainee Policy added and few changes made to TMS

+++ Modules/TMS/Http/Controllers/TraineeController.php
@@ -4,9 +4,11 @@ namespace Modules\TMS\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Routing\Controller;
+use App\Http\Controllers\Controller;
 use Modules\TMS\Services\TraineeService;
 use Modules\TMS\Services\TrainingsService;
+use Modules\TMS\Policies\TraineePolicy;
+use Modules\TMS\Entities\Trainee;
 use Modules\TMS\Http\Requests\TraineeRequest;
 use Illuminate\Support\Facades\Session;
 
@@ -22,6 +24,7 @@ class TraineeController extends Controller
 
     public function index($trainingId = null)
     {
+        $this->authorize('view', Trainee::class);
         $trainees = $trainingId ? $trainees = $this->traineeService->fetchTraineesWithID($trainingId) : null;
 
         $trainings = $this->trainingService->findAll();
@@ -32,6 +35,7 @@ class TraineeController extends Controller
 
     public function create($trainingId)
     {
+        $this->authorize('create', Trainee::class);
         $training = $this->trainingService->findOrFail($trainingId);
         $traineeCount = $this->traineeService->assignedTraineeNo($trainingId);
 
@@ -40,6 +44,7 @@ class TraineeController extends Controller
 
     public function import(Request $request, $trainingId)
     {
+        $this->authorize('create', Trainee::class);
         $training = $this->trainingService->findOrFail($trainingId);
         $traineeCount = $this->traineeService->assignedTraineeNo($trainingId);
 
@@ -57,6 +62,7 @@ class TraineeController extends Controller
 
     public function store(TraineeRequest $request)
     {
+        $this->authorize('create', Trainee::class);
         $storeData = $this->traineeService->save($request->all());
         if($storeData) $msg = __('labels.save_success'); else $msg = __('labels.save_fail');
         Session::flash('message', $msg);
@@ -66,6 +72,7 @@ class TraineeController extends Controller
 
     public function storeImported(Request $request, $training_id)
     {
+        $this->authorize('create', Trainee::class);
         $trainee_mobiles = $request->input('mobile');
 
         $countRow = 0;
@@ -91,6 +98,7 @@ class TraineeController extends Controller
 
     public function show($trainingId)
     {
+        $this->authorize('view', Trainee::class);
         $trainees = $this->traineeService->fetchTraineesWithID($trainingId);
         $training = $this->trainingService->findOrFail($trainingId);
 
@@ -99,6 +107,7 @@ class TraineeController extends Controller
 
     public function edit($traineeId)
     {
+        $this->authorize('update', Trainee::class);
         //$trainee = $this->traineeService->fetchTraineesWithID($traineeId);
         $trainee = $this->traineeService->findOne($traineeId);
         //dd($trainee->training->training_id);
@@ -108,6 +117,7 @@ class TraineeController extends Controller
 
     public function update(TraineeRequest $request, $traineeId)
     {
+        $this->authorize('update', Trainee::class);
         $updateData = array(
             'trainee_first_name' => $request['trainee_first_name'],
             'trainee_last_name' => $request['trainee_last_name'],
@@ -125,6 +135,7 @@ class TraineeController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', Trainee::class);
         $response = $this->traineeService->delete($id);
         if($response) $msg = __('labels.delete_success'); else $msg = __('labels.delete_fail');
         Session::flash('message', $msg);
